Fixing file path issue to run on Windows
The source file name is now extracted from editor.getPath() since, eventually, Atom adds othe informational text to the title, so, it was unsafe to use the window title to infer the file name. All backslashes '\\' characters are replaced by forward slashes prior to split the file path and name, this allow the plugin to properly deal with paths in Windows.